Refactor all tests cases out from the source files to the test folder
